% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfvartest-package.R
\docType{package}
\name{bfvartest-package}
\alias{bfvartest-package}
\alias{bfvartest}
\title{The 'bfvartest' package.}
\description{
Computes Bayes factors for testing (in)equality constrained hypotheses on variances
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
