% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.create_bfvar_object}
\alias{.create_bfvar_object}
\title{Estimates the model using Stan and computes the marginal likelihood}
\usage{
.create_bfvar_object(
  hyp,
  ns,
  sds,
  a,
  compute_ml = TRUE,
  priors_only = FALSE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{hyp}{string specifying the hypothesis}

\item{ns}{vector containing sample sizes}

\item{sds}{vector containing sample standard deviations (with ns - 1 as denominator)}

\item{a}{numeric specifying the value of the parameters of the Dirichlet prior}

\item{compute_ml}{logical specifying whether the marginal likelihood should be computed}

\item{priors_only}{logical specifying whether we should only sample from the prior}

\item{silent}{logical specifying whether to print results from sampling and bridgesampling}

\item{...}{arguments to rstan::sampling}
}
\value{
an object of class 'bfvar', which is a stanfit object with a log marginal likelihood (if desired)
}
\description{
Estimates the model using Stan and computes the marginal likelihood
}
