% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dphi2}
\alias{dphi2}
\title{Computes the posterior density of \phi for the K = 2 case}
\usage{
dphi2(x, n1, n2, sd1, sd2, u = 0.5, logarithm = FALSE)
}
\arguments{
\item{x}{numerical value}

\item{n1}{sample size of group 1}

\item{n2}{sample size of group 2}

\item{sd1}{sample standard deviation of group 1 (with n1 - 1 as denominator)}

\item{sd2}{sample standard deviation of group 2 (with n2 - 1 as denominator)}

\item{u}{parameter of the prior}

\item{logarithm}{a logical specifying whether the log should be taken}
}
\value{
The (log) density at x
}
\description{
Computes the posterior density of \phi for the K = 2 case
}
\examples{

dphi2(seq(0, 2, .01), 100, 100, 1, 1)
}
