% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ksd_test}
\alias{ksd_test}
\title{Computes the k-sample log Bayes factor for all hypotheses}
\usage{
ksd_test(hyp, ns, sds, alpha = 0.5, logarithm = TRUE,
  compute_ml = TRUE, priors_only = FALSE, ...)
}
\arguments{
\item{hyp}{vector of hypotheses}

\item{ns}{vector of sample sizes}

\item{alpha}{prior parameter}

\item{...}{arguments to rstan::sampling}
}
\value{
A list of 'bfvar' objects, which include stanfit objects, log marginal likelihoods, and pairwise Bayes factors
}
\description{
Computes the k-sample log Bayes factor for all hypotheses
}
\examples{
ss <- c(1, 2, 3)
ns <- c(100, 100, 100)
hyp <- c('1=2=3', '1,2,3', '1<2<3')
ksd_test(hyp, ns, ss, 0.50)
}
