% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{one_sample}
\alias{one_sample}
\title{Computes the one-sample log Bayes factor in favour of H1}
\usage{
one_sample(n, s2, popsd, alpha = 0.5)
}
\arguments{
\item{n}{sample size}

\item{s2}{observed sum of squares}

\item{popsd}{population standard deviation we test against}

\item{alpha}{prior parameter}
}
\value{
the one-sample log Bayes factor in favour of H1
}
\description{
Computes the one-sample log Bayes factor in favour of H1
}
\examples{
one_sample(100, 1, 1, 4.50)
}
