% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{two_sample}
\alias{two_sample}
\title{Computes the two-sample log Bayes factor in favour of H1}
\usage{
two_sample(n1, n2, s1, s2, alpha = 0.5, alternative_interval = c(0,
  Inf), null_interval = NULL)
}
\arguments{
\item{n1}{sample size of group 1}

\item{n2}{sample size of group 2}

\item{s1}{observed sum of squares of group 1}

\item{s2}{observed sum of squares of group 2}

\item{alpha}{prior parameter}

\item{alternative_interval}{interval for the alternative hypothesis (e.g., c(1, Inf) and c(0, 1) give directed tests)}

\item{null_interval}{interval for the null hypothesis (e.g., c(0.9, 1.1))}
}
\value{
the two-sample log Bayes factor in favour of H1
}
\description{
Computes the two-sample log Bayes factor in favour of H1
}
\examples{
two_sample(100, 200, 1, 2, 4.5, alternative_interval = c(1, Inf)) # one-sided test
two_sample(100, 200, 1, 2, 4.5, alternative_interval = c(0, 1)) # one-sided test
two_sample(100, 200, 1, 2, 4.5, null_interval = c(0.9, 1.1)) # interval Bayes factor
two_sample(100, 200, 1, 2, 4.5, alternative_interval = c(1, Inf), null_interval = c(0.9, 1.1)) # one-sided interval Bayes factor
}
