% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.create_bfvar_object}
\alias{.create_bfvar_object}
\title{Estimates the model using Stan and computes the marginal likelihood}
\usage{
.create_bfvar_object(hyp, ns, ss, a, compute_ml = TRUE,
  priors_only = FALSE, ...)
}
\description{
Estimates the model using Stan and computes the marginal likelihood
}
