% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.create_bfvar_object}
\alias{.create_bfvar_object}
\title{Estimates the model using Stan and computes the marginal likelihood}
\usage{
.create_bfvar_object(
  hypothesis,
  ns,
  ss,
  a,
  compute_ml = TRUE,
  priors_only = FALSE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{hypothesis}{a string specifying the hypothesis}

\item{ns}{a vector containing sample sizes}

\item{ss}{a vector containing sample sum of squares}

\item{a}{a vector specifying the value of the parameters of the Dirichlet prior}

\item{compute_ml}{a logical specifying whether the marginal likelihood should be computed}

\item{priors_only}{a logical specifying whether we should only sample from the prior}

\item{silent}{a logical specifying whether to print results from sampling and bridgesampling}

\item{...}{arguments to rstan::sampling}
}
\value{
an object of class 'bfvar', which is a stanfit object with a log marginal likelihood (if desired)
}
\description{
Estimates the model using Stan and computes the marginal likelihood
}
